Updated Persona to Persona function
Fixed the final result persona not working because the if statements were poorly written. Updated the if statement to work and turned the for loop into a while loop. Did it in order to remove break statements that I had in there. I joined up the if statements as well to make it more clear.